### 2022-10-07 #556 @SebKrantz
dfms: Efficient Estimation of Dynamic Factor Models for R
Submitting Author Name: Sebastian Krantz
Submitting Author Github Handle: @SebKrantz
Other Package Authors Github handles: @rbagd
Repository:  https://github.com/SebKrantz/dfms
Version submitted: 0.1.2
Submission type: Stats
Badge grade: bronze

Version accepted: TBD
---

```
Package: dfms
Version: 0.1.2
Title: Dynamic Factor Models
Authors@R: c(person("Sebastian", "Krantz", role = c("aut", "cre"), email = "sebastian.krantz@graduateinstitute.ch"),
             person("Rytis", "Bagdziunas", role = "aut"))
Description: Efficient estimation of Dynamic Factor Models using the Expectation Maximization (EM) algorithm
  or Two-Step (2S) estimation, on datasets with missing data. The implementation follows advances in the econometric
  literature: estimation can be done either by running the Kalman Filter and Smoother once with initial values
  from PCA - following Doz, Giannone and Reichlin (2011) (2S) - or via iterated Kalman Filtering and Smoothing until EM
  convergence - following Doz, Giannone and Reichlin (2012) - or using the adapted EM algorithm of Banbura and Modugno
  (2014), allowing estimation with arbitrary patterns of missing data. The implementation makes heavy use of the
  Armadillo C++ library and the collapse package, providing for particularly speedy estimation. A comprehensive set of
  methods supports interpretation/visualization of the model and forecasting. Information criteria to choose the number
  of factors are also provided - following Bai and Ng (2002).
  --- Key References: ---
  Doz, C., Giannone, D., & Reichlin, L. (2011). A two-step estimator for large approximate dynamic
       factor models based on Kalman filtering. Journal of Econometrics, 164(1), 188-205.
  Doz, C., Giannone, D., & Reichlin, L. (2012). A quasi-maximum likelihood approach for large, approximate
       dynamic factor models. Review of Economics and Statistics, 94(4), 1014-1024.
  Banbura, M., & Modugno, M. (2014). Maximum likelihood estimation of factor models on datasets with arbitrary
       pattern of missing data. Journal of Applied Econometrics, 29(1), 133-160.
Imports: Rcpp (>= 1.0.1), collapse (>= 1.8.0)
LinkingTo: Rcpp, RcppArmadillo
    xts,
    vars,
    covr
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one appropriate box below):

- [x] A pre-submission inquiry has been approved in [issue #555](https://github.com/ropensci/software-review/issues/555)

Anybody working with time series. The package is useful for dimensionality reduction and forecasting with a large amount of time series.

    - *The first implementation of a novel algorithm*; or
    - *The first implementation within* **R** *of an algorithm which has previously been implemented in other languages or contexts*; or
    - *An improvement on other implementations of similar algorithms in* **R**.

See README.md, *dfms* implements simple baseline versions of algorithms that have been around for a while in Matlab, and in other langaues (R, Python, Julia), but inside more elaborate nowcasting codes - thus not directly accessible, and less efficient. It is the only *pure baseline* implementation of the algorithms proposed by the 3 academic references mentioned in the description available for R and ready for CRAN.

    Please include hyperlinked references to all other relevant software.

The software is actually a reboot and massive improvement upon [dynfactoR](https://github.com/rbagd/dynfactoR), an abandoned software project. Generalizations of the functionality are provided by [nowcasting](<https://github.com/nmecsys/nowcasting>) and [nowcastDFM](<https://github.com/dhopp1/nowcastDFM>), which fit dynamic factor models specific to mixed-frequency nowcasting applications. These packages are currently not on CRAN (they were archived) and also not very well maintained. Package [MARSS](<https://CRAN.R-project.org/package=MARSS>) can be used to fit dynamic factor models, but has a complicated API and fails on bigger datasets. The only really useful and well maintained dynamic factor modelling package for R is [bayesdfa](https://github.com/fate-ewi/bayesdfa), which is also on CRAN, and fits bayesian dynamic factor models with Stan. I expect *dfms* to provide substantially faster estimation than *bayesdfa*.  There are various other codes for Python and Julia on GitHub, including an implementation in the popular [statsmodels](https://www.statsmodels.org/dev/examples/notebooks/generated/statespace_dfm_coincident.html#examples-notebooks-generated-statespace-dfm-coincident--page-root) library, but I did not engage with those as my primary tool remains R and I wanted to create an efficient baseline implementation for R that follows advances in the econometrics literature (PCA + EM Algorithm based estimation).

Bronze

There are still some `autotest` issues, especially for the main `DFM()` function, bit I do not understand those as all inputs received the maximum extent of checking. See [lines 211-226](https://github.com/SebKrantz/dfms/blob/main/R/DFM.R).  I also don't understand the note in `pkgcheck` requesting CI checks. The package receives CI through GitHub Actions (all plattforms) and test coverage is uploaded to codecov.io.
