### 2022-10-10 455e396
Implementing CRAN comments: DOIs instead of full references.
