+++ DESCRIPTION
@@ -4,34 +4,21 @@ Title: Dynamic Factor Models
 Authors@R: c(person("Sebastian", "Krantz", role = c("aut", "cre"), email = "sebastian.krantz@graduateinstitute.ch"),
              person("Rytis", "Bagdziunas", role = "aut"))
 Description: Efficient estimation of Dynamic Factor Models using the Expectation Maximization (EM) algorithm 
-  or Two-Step (2S) estimation, on datasets with missing data. The implementation follows advances in the econometric 
-  literature: estimation can be done either by running the Kalman Filter and Smoother once with initial values 
-  from PCA - following Doz, Giannone and Reichlin (2011) (2S) - or via iterated Kalman Filtering and Smoothing until EM 
-  convergence - following Doz, Giannone and Reichlin (2012) - or using the adapted EM algorithm of Banbura and Modugno 
-  (2014), allowing estimation with arbitrary patterns of missing data. The implementation makes heavy use of the 
-  Armadillo C++ library and the collapse package, providing for particularly speedy estimation. A comprehensive set of 
-  methods supports interpretation/visualization of the model and forecasting. Information criteria to choose the number 
-  of factors are also provided - following Bai and Ng (2002).
-  --- Key References: ---
-  Doz, C., Giannone, D., & Reichlin, L. (2011). A two-step estimator for large approximate dynamic 
-       factor models based on Kalman filtering. Journal of Econometrics, 164(1), 188-205.
-  Doz, C., Giannone, D., & Reichlin, L. (2012). A quasi-maximum likelihood approach for large, approximate 
-       dynamic factor models. Review of Economics and Statistics, 94(4), 1014-1024.
-  Banbura, M., & Modugno, M. (2014). Maximum likelihood estimation of factor models on datasets with arbitrary 
-       pattern of missing data. Journal of Applied Econometrics, 29(1), 133-160.
+  or Two-Step (2S) estimation, supporting datasets with missing data. The implementation follows advances in the 
+  econometric literature: estimation can be done either by running the Kalman Filter and Smoother once with initial 
+  values from PCA - 2S estimation following Doz, Giannone and Reichlin (2011) <doi:10.1016/j.jeconom.2011.02.012> - 
+  or via iterated Kalman Filtering and Smoothing until EM convergence - following Doz, Giannone and Reichlin (2012) 
+  <doi:10.1162/REST_a_00225> - or using the adapted EM algorithm of Banbura and Modugno (2014) <doi:10.1002/jae.2306>, 
+  allowing arbitrary patterns of missing data. The implementation makes heavy use of the 'Armadillo' 'C++' library and 
+  the 'collapse' package, providing for particularly speedy estimation. A comprehensive set of methods supports 
+  interpretation and visualization of the model as well as forecasting. Information criteria to choose the number 
+  of factors are also provided - following Bai and Ng (2002) <doi:10.1111/1468-0262.00273>.
 URL: https://sebkrantz.github.io/dfms/
 BugReports: https://github.com/SebKrantz/dfms/issues
 Depends: R (>= 3.0.0)
 Imports: Rcpp (>= 1.0.1), collapse (>= 1.8.0)
 LinkingTo: Rcpp, RcppArmadillo
-Suggests: 
-    xts,
-    vars,
-    magrittr,
-    testthat (>= 3.0.0),
-    knitr,
-    rmarkdown,
-    covr
+Suggests: xts, vars, magrittr, testthat (>= 3.0.0), knitr, rmarkdown, covr
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
