### 2022-10-10 1fd7415
Implementing CRAN comments: not using T and F as variable names.
