+++ R/methods.R
@@ -36,12 +36,12 @@ print.dfm <- function(x, digits = 4L, ...) {
 summary.dfm <- function(object, method = switch(object$em.method, none = "2s", "qml"), ...) {
 
   X <- object$X_imp
-  F <- switch(tolower(method), pca = object$F_pca, `2s` = object$F_2s, qml = object$F_qml, stop("Unkown method", method))
+  Fa <- switch(tolower(method), pca = object$F_pca, `2s` = object$F_2s, qml = object$F_qml, stop("Unkown method", method))
   A <- object$A
   r <- dim(A)[1L]
   p <- dim(A)[2L] / r
   C <- object$C
-  res <- X - tcrossprod(F, C)
+  res <- X - tcrossprod(Fa, C)
   anymissing <- object$anyNA
   if(anymissing) res[attr(X, "missing")] <- NA
   rescov <- pwcov(res, use = if(anymissing) "pairwise.complete.obs" else "everything", P = TRUE)
@@ -50,9 +50,9 @@ summary.dfm <- function(object, method = switch(object$em.method, none = "2s", "
   summ <- list(info = c(n = dim(X)[2L], T = dim(X)[1L], r = r, p = p,
                         `%NA` = if(anymissing) sum(attr(X, "missing")) / prod(dim(X)) * 100 else 0),
                call = object$call,
-               F_stats = msum(F),
+               F_stats = msum(Fa),
                A = A,
-               F_cov = pwcov(F, P = TRUE),
+               F_cov = pwcov(Fa, P = TRUE),
                Q = object$Q,
                C = C,
                R_diag = diag(object$R),
@@ -142,30 +142,30 @@ plot.dfm <- function(x,
                      method = switch(x$em.method, none = "2s", "qml"),
                      type = c("joint", "individual", "residual"),
                      scale.factors = TRUE, ...) {
-  F <- switch(tolower(method[1L]),
+  Fa <- switch(tolower(method[1L]),
               all = cbind(x$F_pca, setColnames(x$F_2s, paste("TwoStep", colnames(x$F_2s))),
                           if(length(x$F_qml)) setColnames(x$F_qml, paste("QML", colnames(x$F_qml))) else NULL),
               pca = x$F_pca, `2s` = x$F_2s, qml = x$F_qml, stop("Unknown method:", method[1L]))
 
-  nf <- dim(F)[2L]
+  nf <- dim(Fa)[2L]
   allests <- tolower(method[1L]) == "all"
   dots <- list(...)
 
   switch(tolower(type[1L]),
     joint = {
       Xr <- frange(x$X_imp)
-      if(scale.factors) F <- fscale(F)
-      Fr <- frange(F)
+      if(scale.factors) Fa <- fscale(Fa)
+      Fr <- frange(Fa)
       ts.plot(x$X_imp, col = "grey85", ylim = c(min(Xr[1L], Fr[1L]), max(Xr[2L], Fr[2L])),
               ylab = if(is.null(dots$ylab)) "Value" else dots$ylab,
               main = if(is.null(dots$main)) "Standardized Series and Factor Estimates" else dots$main, ...)
       cols <- rainbow(nf)
-      for (i in seq_len(nf)) lines(F[, i], col = cols[i])
-      legend("topleft", colnames(F), col = cols, lty = 1, bty = "n", ncol = if(allests) 3L else 1L)
+      for (i in seq_len(nf)) lines(Fa[, i], col = cols[i])
+      legend("topleft", colnames(Fa), col = cols, lty = 1, bty = "n", ncol = if(allests) 3L else 1L)
[…]
