### 2022-10-10 4eff27c
Implementing CRAN comments: resetting par().
