### 2022-10-10 aff9415
Reverting the redundant on.exit statement.
