### 2022-10-27 b7f0116
Documentation minors, and saving loglik_all without growing vector all the time.
