### 2023-01-03 #556 @santikka
---

### Compliance with Standards

- [ ] This package complies with a sufficient number of standards for a (bronze/silver/gold) badge
- [ ] This grade of badge is the same as what the authors wanted to achieve

The following standards currently deemed non-applicable (through tags of `@srrstatsNA`) could potentially be applied to future versions of this software:

I'm currently unable to assess how well this package conforms to the standards for two reasons. First, all of the `@srrstats` tags are still included in the `srr-stats-standards.R` file and are not placed in appropriate places in the code. Second, the compliance of only a handful of the standards has been explained.

---

### General Review

- [ ] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING
- [x] The documentation is sufficient to enable general use of the package beyond one specific use case

#### Algorithms

As I understand it, the one of the main advantages of this package is computational speed, making C++ a valid choice for implementing the core algorithms. I would like to see some tests to substantiate the performance claims of this package vs. packages such as `MARSS` and `bayesdfa` that were mentioned by the authors.

While the package has impressive test coverage, I found the suite of tests lacking especially with regards to the inputs for the exported functions. There are several instances where the inputs are not thoroughly checked, for example:
```R
.VAR(diff(EuStockMarkets), -1)
 Error in `[.default`(x, (p + 1L - i):(TT - i), ) :
only 0's may be mixed with negative subscripts
```
I'm also not convinced that the results of `autotest` have been properly accounted for. At the time of writing this review, running `autotest` on the package produces a data frame with 200 rows.

#### Visualisation (where appropriate)

Visualisations are mostly clear and aid the primary purposes of statistical interpretation of the results. I don't think there is a risk of statistical misinterpretation.

There is a small issue where the plot legend may overlap with the plotted values (for example when running `plot(mod, type = "individual", method = "all")` in the examples of `plot.dfm`), perhaps the authors could move the legend outside of the main figure area.

#### Package Design

Overall, the package seems mostly well designed for its intended purpose. The code has been thoroughly annotated making it easy to understand, although some of the lines are very long, and the indentation is sometimes inconsistent.

There are some curious design choices, like using `switch` instead of `match.arg` to check function arguments. Also, this part in `DFM.R` puzzled me:
```R
# Quoting some functions that need to be evaluated iteratively
.EM_DGR <- quote(EMstepDGR(X, A, C, Q, R, F_0, P_0, cpX, n, r, sr, TT, rQi, rRi))
.EM_BM <- quote(EMstepBMOPT(X, A, C, Q, R, F_0, P_0, XW0, W, n, r, sr, TT, dgind, dnkron, dnkron_ind, rQi, rRi))
.KFS <- quote(SKFS(X, A, C, Q, R, F_0, P_0))
```
These expressions are then `eval`'d later, why exactly is this done as opposed to simply calling these functions directly?

I'm also curious as to what exactly is the logic in choosing which parts of the code are written in R and which in C++. It seems to me that there is quite a lot of matrix algebra going on in the R code. I would have expected that R would only be used to check and prepare the inputs for C++, and to provide outputs such as plots.

Print methods should return their respective argument objects invisibly (i.e., for `dfm`, `dfm_summary`, `dfm_forecast`, and `ICr`).

Since this is a non-tidyverse package, it is limited in terms of inter-operability in relation to other packages. For example, only base R graphics are available.

---

This package does not conform to the guidelines in at least the following aspects:

  * Inconsistent use of assignment operators
  * No top-level documentation
  * No CITATION file
  * README should also show the results of the usage example (I suggest using .Rmd)
