### 2023-01-12 3828747
Better solution.
