### 2023-01-12 799dd43
Final solution.
