### 2023-02-08 #556 @eeholmes
Sorry, in my comment I meant the EM algorithm in the **MARSS** package struggles (= very slow) with the factors loading (C matrix) estimation. I wasn't referring to to the EM algorithm in Banbura, M., & Modugno, M. (2014) and Doz, C., Giannone, D., & Reichlin, L. (2012).
