### 2023-02-08 #556 @eeholmes
@SebKrantz Can I get a little help on something?

For identifiability, the C matrix must be constrained as Banbura & Modugno discuss here on the top of page 14
https://www.ecb.europa.eu/pub/pdf/scpwps/ecbwp1189.pdf

But when I fit a DFA and look at the C matrix
```
model1 <- DFA(...)
model1$C
```

The C matrix doesn't have the identifiability constraint. So you must have done a rotation. Can you tell me what rotation DFA() is doing? I am trying to get the original constrained C matrix used in your EM algorithm. This C (which they call Lambda):

Thanks!
