### 2023-02-08 #556 @SebKrantz
@eeholmes, I am using a M1 mac and I have no problems building the package, binaries for mac are also available on CRAN. So the algorithm is a direct translation of B&M's original Matlab codes, in particular the standard single-frequency DFM estimation without autocorrelated errors forund in [this file](https://github.com/SebKrantz/dfms/blob/main/misc/BM2014/RepFilesBanburaModugnoJAE14/functions/ML/EM_DFM_SS.m). I did a 1:1 translation of the code to R [here](https://github.com/SebKrantz/dfms/tree/main/misc/BM2014/BM14_R/EM_DFM_SS) and optimitzed this implementation in R [here](https://github.com/SebKrantz/dfms/tree/main/misc/BM2014/BM14_R/EM_DFM_SS_OPT). This was then integrated with the C++ Kalman Filters and Smoothers to produce the implementation available in `dfms::DFM()`.

Regarding the rotation, the Kalman filter and smoother was initialized with values from PCA, so I guess that gives orthogonal initial factor estimates and identifies the matrix. It may also be that B&M provide comments pertaining to mixed frequency estimation which I haven't implemented yet because as of now dfms estimates single frequency DFM's without autocorrelated errors.
