### 2023-02-16 1c50d3f
Adding option na.keep to fitted.dfm: fixes #45.
