+++ NEWS.md
@@ -1,3 +1,7 @@
+# dfms 0.1.5
+
+* Added argument `na.keep = TRUE` to `fitted.dfm`. Setting `na.keep = FALSE` allows interpolation of data based on the DFM. Thanks @apoorvalal (#45).
+
 # dfms 0.1.4
 
 * Fixed minor bug in `summary.dfm` occurring if only one factor was estimated (basically an issue with dropping matrix dimensions which lead the factor summary statistics to be displayed without names).
