### 2023-02-16 b07bdf6
Adding r-universe badge and adjusting installation instructions.
