### 2023-02-23 bfc3f22
Starting a more theoretical vignette.
