+++ misc/dynamic_factor_models/dynamic_factor_models.tex
@@ -1,0 +1,190 @@
+\documentclass[a4paper]{article}
+
+%% Language and font encodings
+\usepackage[english]{babel}
+\usepackage[utf8x]{inputenc}
+% \usepackage[T1]{fontenc}
+
+%% Sets page size and margins
+\usepackage[a4paper,top=3cm,bottom=2cm,left=3cm,right=3cm,marginparwidth=1.75cm]{geometry}
+
+%% Useful packages
+\usepackage{amsmath}
+\usepackage{graphicx}
+\usepackage[colorlinks=true, allcolors=blue]{hyperref}
+\usepackage{apacite} % APA style citations
+\AtBeginDocument{\urlstyle{APACsame}}  % Links in APA citytions same formatting
+\usepackage{tabulary}
+\usepackage{natbib} % natbib citations: \citep{} and \citet{} for in-text
+\usepackage{booktabs}
+% If you use \tableofcontents, this adjusts the name
+%\addto\captionsenglish{\renewcommand*\contentsname{Table of Contents}}
+
+
+\title{\textbf{Dynamic Factor Models}\\ A Short Introduction}
+\author{Sebastian Krantz}
+
+\begin{document}
+\maketitle
+
+Dynamic factor models (DFM) postulate that a small number of unobserved "factors" can be used to explain a substantial portion of the variation and dynamics in a larger number of observed time series. In addition to producing estimates of the unobserved factors, dynamic factor models have many uses in forecasting and macroeconomic monitoring. One popular application for these models is "nowcasting", in which higher-frequency data is used to produce "nowcasts" of series that are only published at a lower frequency (Fulton, 2020). \newline
+
+Dynamic Factor Models are set up in State-Space form and can be estimated using the Kalman Filter and a number of solution algorithms, the most popular one in the economics literature being the Expectation Maximization (EM) algorithm, due to it's robust numerical properties and the mixed frequency modification of \citet{banbura2014maximum}. \newline
+
+\section{The Canonical (Exact) DFM}
+
+A baseline dynamic factor model can be written as follows
+
+\begin{align} \label{eq:do}
+\textbf{x}_t &= \textbf{C}_0 \textbf{f}_t + \textbf{e}_t, \qquad\qquad\, \textbf{e}_t\sim N(\textbf{0}, \textbf{R}) \\ \label{eq:dt}
+\textbf{f}_t &= \sum_{j=1}^p \textbf{A}_j \textbf{f}_{t-j} + \textbf{u}_t, \quad\ \textbf{u}_t\sim  N(\textbf{0}, \textbf{Q}_0),
+\end{align}
+
+where the unobserved factors $\textbf{f}_t$ are evolve according to a VAR(p) process and  \newline
+
+\begin{tabular}{p{1cm}p{12cm}}
+$\textbf{x}_t$ & $n \times 1$ vector of observed series at time $t$: $(x_{1t}, \dots, x_{nt})'$. Some observations can be missing.  \\\\
+$\textbf{f}_t$ & $r \times 1$ vector of factors at time $t$: $(f_{1t}, \dots, f_{rt})'$.\\\\
+$\textbf{C}_0$ & $n \times r$ measurement (observation) matrix.\\\\
+$\textbf{A}_j$ & $r \times r$ state transition matrix at lag $j$. \\\\
[…]
