+++ misc/dynamic_factor_models/dynamic_factor_models.tex
@@ -19,15 +19,16 @@
 \usepackage{booktabs}
 % If you use \tableofcontents, this adjusts the name
 %\addto\captionsenglish{\renewcommand*\contentsname{Table of Contents}}
+\setcounter{MaxMatrixCols}{20} % https://tex.stackexchange.com/questions/436646/matrix-extra-alignment-tab-has-been-changed-to-cr
 
 
-\title{\textbf{Dynamic Factor Models}\\ A Short Introduction}
+\title{\textbf{Dynamic Factor Models}\\ A Very Short Introduction}
 \author{Sebastian Krantz}
 
 \begin{document}
 \maketitle
 
-Dynamic factor models (DFM) postulate that a small number of unobserved "factors" can be used to explain a substantial portion of the variation and dynamics in a larger number of observed time series. In addition to producing estimates of the unobserved factors, dynamic factor models have many uses in forecasting and macroeconomic monitoring. One popular application for these models is "nowcasting", in which higher-frequency data is used to produce "nowcasts" of series that are only published at a lower frequency (Fulton, 2020). \newline
+Dynamic factor models (DFM) postulate that a small number of unobserved "factors" can be used to explain a substantial portion of the variation and dynamics in a larger number of observed time series. In addition to producing estimates of the unobserved factors, dynamic factor models have many uses in forecasting and macroeconomic monitoring. One popular application for these models is "nowcasting", in which higher-frequency data is used to produce "nowcasts" of series that are only published at a lower frequency (Chad Fulton, 2020). \newline
 
 Dynamic Factor Models are set up in State-Space form and can be estimated using the Kalman Filter and a number of solution algorithms, the most popular one in the economics literature being the Expectation Maximization (EM) algorithm, due to it's robust numerical properties and the mixed frequency modification of \citet{banbura2014maximum}. \newline
 
@@ -93,7 +94,7 @@ Particularly assumption 4 is quite restrictive since it stipulates that all time
 \textbf{x}_t = \sum_{i=0}^q \textbf{C}_i \textbf{f}_{t-i} + \textbf{e}_t, \qquad \textbf{e}_t\sim N(\textbf{0}, \textbf{R}).
 \end{equation}
 In this case the stacked notation remains the same as long as $q < r-1$, with observation matrix 
-\begin{equation}
+\begin{equation} \label{eq:Clags}
 \textbf{C}_{(n \times rp)}  = (\textbf{C}_0, \textbf{C}_1, \dots, \textbf{C}_q, \textbf{0}, \dots, \textbf{0})
 \end{equation}
[…]
