+++ misc/dynamic_factor_models/dynamic_factor_models.tex
@@ -13,7 +13,7 @@
 \usepackage{graphicx}
 \usepackage[colorlinks=true, allcolors=blue]{hyperref}
 \usepackage{apacite} % APA style citations
-\AtBeginDocument{\urlstyle{APACsame}}  % Links in APA citytions same formatting
+\AtBeginDocument{\urlstyle{APACsame}}  % Links in APA citations same formatting
 \usepackage{tabulary}
 \usepackage{natbib} % natbib citations: \citep{} and \citet{} for in-text
 \usepackage{booktabs}
@@ -28,40 +28,39 @@
 \begin{document}
 \maketitle
 
-Dynamic factor models (DFM) postulate that a small number of unobserved "factors" can be used to explain a substantial portion of the variation and dynamics in a larger number of observed time series. In addition to producing estimates of the unobserved factors, dynamic factor models have many uses in forecasting and macroeconomic monitoring. One popular application for these models is "nowcasting", in which higher-frequency data is used to produce "nowcasts" of series that are only published at a lower frequency (Chad Fulton, 2020). \newline
+Dynamic factor models (DFMs) postulate that a small number of latent factors explain the common dynamics of a larger number of observed time series \citep{stock2016dynamic}. DFMs provide estimates of these unobserved factors and their joint dynamics, with many applications in forecasting, timer series interpolation, and macroeconomic monitoring, such as the creation of coincident business cycle indicators \citep{mariano2003new}. Another popular application is "nowcasting", where large volumes of time series data released at higher frequencies are synthesized to produce real-time estimates of low-frequency leading indicators such as GDP \citep{bok2018macroeconomic}. \newline
 
-Dynamic Factor Models are set up in State-Space form and can be estimated using the Kalman Filter and a number of solution algorithms, the most popular one in the economics literature being the Expectation Maximization (EM) algorithm, due to it's robust numerical properties and the mixed frequency modification of \citet{banbura2014maximum}. \newline
+DFMs are set up in State Space form and can be estimated using the Kalman Filter and several solution algorithms, the most popular one in the economics literature being the Expectation Maximization (EM) algorithm, due to its robust numerical properties \citep{doz2012quasi} and the popular mixed frequency generalization of \citet{banbura2014maximum}. % \newline
 
 \section{The Canonical (Exact) DFM}
 
-A baseline dynamic factor model can be written as follows
-
+A canonical baseline dynamic factor model can be written as
 \begin{align} \label{eq:do}
 \textbf{x}_t &= \textbf{C}_0 \textbf{f}_t + \textbf{e}_t, \qquad\qquad\, \textbf{e}_t\sim N(\textbf{0}, \textbf{R}) \\ \label{eq:dt}
 \textbf{f}_t &= \sum_{j=1}^p \textbf{A}_j \textbf{f}_{t-j} + \textbf{u}_t, \quad\ \textbf{u}_t\sim  N(\textbf{0}, \textbf{Q}_0),
 \end{align}
-
-where the unobserved factors $\textbf{f}_t$ are evolve according to a VAR(p) process and  \newline
[…]
