+++ misc/BM2014/BM14_Julia/EM_DFM_SS_idio/EM_DFM_SS_idio.jl
@@ -1,0 +1,141 @@
+###########################################
+# Julia Implementation of Single Frequency
+# DFM with AR(1) Observation Residuals
+###########################################
+
+# Installation and loading
+# using Pkg
+# Pkg.add("XLSX")
+# Pkg.add("NaNStatistics") # nanmean etc.
+# Pkg.add("DSP") # filt()
+# Pkg.add("DataInterpolations") # CubicSpline()
+# Pkg.add("BlockDiagonals") # BlockDiagonal()
+using XLSX
+using Statistics
+using NaNStatistics
+using DSP
+using DataInterpolations
+using LinearAlgebra
+using BlockDiagonals
+
+eye(n) = Matrix{Float64}(I, n, n)
+
+# Parameters
+P = (r = 5, p = 2, i_idio = repeat([1], size(X, 2)), max_iter = 100)
+
+mutable struct Nanopt
+    method
+    k
+end
+
+# Functions 
+include("runKF.jl")
+include("remNaNs_spline.jl")
+include("Procedures.jl")
+
+
+function EM_DFM_SS_idio(X, P)
+
+    # X: T x N panel of data
+    # P: structure containing settings of the model and estimation 
+
+    thresh = 1e-4
+    r = P.r # number of factors
+    p = P.p # number of lags in the factor VAR
+    max_iter = P.max_iter # maximal number of EM iterations
+    i_idio = P.i_idio # logical vector Nx1 with 1's for variables with serially correlated idio
+
+    #--------------------------------------------------------------------------
+    # Preparation of the data
+    #--------------------------------------------------------------------------
+    T, N = size(X)
+
+    # Standardise x
+    Mx = nanmean(X, dims = 1)
+    Wx = nanstd(X, dims = 1)
+
+    # Standardize X using Mx and Wx  
+    xNaN = (X .- Mx) ./ Wx
+    # xNaN = X;
+
+    #--------------------------------------------------------------------------
+    # Initial Conditions
+    #--------------------------------------------------------------------------
+
+    #Removing missing values (for initial estimators)
+
+    optNaN = Nanopt(2, 3) # Remove leading and closing zeros
+
+    A, C, Q, R, Z_0, V_0 = InitCond(xNaN, r, p, optNaN, i_idio)
+
+    # some auxiliary variables for the iterations
+    previous_loglik = -Inf
+    num_iter = 0
+    LL = -Inf
+    converged = false
+
+    # y for the estimation is WITH missing data
+    y = xNaN'
+
+    #--------------------------------------------------------------------------
+    #THE EM LOOP
+    #--------------------------------------------------------------------------
+
+    #The model can be written as
+    #y = C*Z + e;
+    #Z = A*Z(-1) + v
+    #where y is NxT, Z is (pr)xT, etc
+
+    #remove the leading and ending nans for the estimation
+    optNaN.method = 3
+    y_est = remNaNs_spline(xNaN, optNaN)[1]'
+    wnidio = findall(i_idio .== 0)
+
+    decrease = zeros(max_iter+1)
+    LL = zeros(max_iter+1)
+
+    while num_iter < max_iter && !converged 
+        C_new, R_new, A_new, Q_new, loglik, Z_0, V_0 = EMstep(y_est, A, C, Q, R, Z_0, V_0, r, p, i_idio) 
+        
+        C = C_new
+        R[wnidio, wnidio] .= R_new[wnidio, wnidio]
+        A = A_new
+        Q = Q_new
+
+        if num_iter > 2
+            converged, decrease[num_iter+1] = em_converged(loglik, previous_loglik, thresh, true)
[…]
