### 2023-03-25 d7b8872
Adding baseline BM14 code for AR(1) errors.
