+++ R/EMBM_idio.R
@@ -1,0 +1,134 @@
+
+# Define the inputs of the function
+# X: a matrix with n rows and TT columns representing the observed data at each time t.
+# A: a transition matrix with size r*p+r*r.
+# C: an observation matrix with size n*r.
+# Q: a covariance matrix for the state equation disturbances with size r*p+r*r.
+# R: a covariance matrix for the observation disturbances with size n*n.
+# Z_0: the initial value of the state variable.
+# V_0: the initial value of the variance-covariance matrix of the state variable.
+# r: order of the state autoregression (AR) process.
+# p: number of lags in the state AR process.
+# i_idio: a vector of indicators specifying which variables are idiosyncratic noise variables.'
+
+# Z_0 = F_0; V_0 = P_0
+EMstepBMidio = function(X, A, C, Q, R, Z_0, V_0, XW0, W, dgind, dnkron, dnkron_ind, r, p) {
+
+  #Define dimensions of the input arguments
+  TT = dim(X)[1L]
+  n = dim(X)[2L]
+  sr = seq_len(r)
+  srp = seq_len(r*p)
+  rp1nr = (r*p+1L):ncol(A)
+
+  # Compute expected sufficient statistics for a single Kalman filter sequence
+  # Run the Kalman filter with the current parameter estimates
+  kfs_res = SKFS(X, A, C, Q, R, Z_0, V_0, TRUE)
+
+  # Extract values from the Kalman Filter output
+  Zsmooth = kfs_res$F_smooth
+  Vsmooth = kfs_res$P_smooth
+  VVsmooth = kfs_res$PPm_smooth
+  Zsmooth0 = kfs_res$F_smooth_0
+  Vsmooth0 = kfs_res$P_smooth_0
+  loglik = kfs_res$loglik
+
+  # Perform algebraic operations involving E(Z_t), E(Z_{t-1}), Y_t, Y_{t-1} to update A_new, Q_new, C_new, and R_new
+
+  # Normal Expectations to get Z(t+1) and A(t+1).
+  tmp = rbind(Zsmooth0[srp], Zsmooth[-TT, srp, drop = FALSE])
+  tmp2 = sum3(Vsmooth[srp, srp, -TT, drop = FALSE])
+  EZZ = crossprod(Zsmooth[, srp, drop = FALSE]) %+=% (tmp2 + Vsmooth[srp, srp, TT])                   # E(Z'Z)
+  EZZ_BB = crossprod(tmp) %+=% (tmp2 + Vsmooth0[srp, srp])                                            # E(Z(-1)'Z(-1))
+  EZZ_FB = crossprod(Zsmooth[, srp, drop = FALSE], tmp) %+=% sum3(VVsmooth[srp, srp,, drop = FALSE])  # E(Z'Z(-1))
+
+  # Expectations on the erros (u_t).
+  tmp = rbind(Zsmooth0[rp1nr], Zsmooth[-TT, rp1nr, drop = FALSE])
+  tmp2 = sum3(Vsmooth[rp1nr, rp1nr, -TT, drop = FALSE])
+  EZZ_u = diag(crossprod(Zsmooth[, rp1nr, drop = FALSE])) + diag(tmp2 + Vsmooth[rp1nr, rp1nr, TT])                     # E(Z'Z)
[…]
