+++ NAMESPACE
@@ -27,8 +27,10 @@ export(em_converged)
 export(tsnarmimp)
 importFrom(collapse,"%*=%")
 importFrom(collapse,"%+=%")
+importFrom(collapse,"%-=%")
 importFrom(collapse,"%/=%")
 importFrom(collapse,"%=%")
+importFrom(collapse,"%r*%")
 importFrom(collapse,TRA.matrix)
 importFrom(collapse,ckmatch)
 importFrom(collapse,fmedian)
