### 2023-03-27 b5b0ba1
Adding additional return values and adjusting documentation a bit.
