### 2023-03-27 9c09f49
Adding tolerance to init_cond_idio_ari: for R diagonal.
