### 2023-03-27 41201ff
Adding covariance restrictions for estimation with AR(1) errors.
