### 2023-03-27 fe9b0cd
Further simplification.
