+++ R/EMBM_idio.R
@@ -58,7 +58,7 @@ EMstepBMidio = function(X, A, C, Q, R, Z_0, V_0, XW0, W, dgind, dnkron, dnkron_i
   } else Q_new[sr, sr] = diag(r)
 
   # Errors
-  A_new[rp1nr, rp1nr] = EZZ_FB_u * (1/EZZ_BB_u) # %r*% # Same as EZZ_FB_u is diagonal...
+  A_new[rp1nr, rp1nr] = EZZ_FB_u / EZZ_BB_u  # %r*% (1/EZZ_BB_u) # Same as EZZ_FB_u is diagonal...
   if(rRi) {
     if(rRi == 2L) stop("Cannot estimate unrestricted observation covariance matrix together with AR(1) serial correlation")
     Q_new[rp1nr, rp1nr] = (diag(EZZ_u) - A_new[rp1nr, rp1nr] * EZZ_FB_u) / TT # tcrossprod(A_new[rp1nr, rp1nr, drop = FALSE], EZZ_FB_u) # Same as EZZ_FB_u is diagonal...
