+++ misc/BM2014/BM14_Julia/KalmanFiltering.jl
@@ -1,0 +1,156 @@
+using LinearAlgebra
+
+# Kalman Filter and Smoother
+
+# X Data matrix (T x n)
+# A Transition matrix (rp x rp)
+# C Observation matrix (n x rp)
+# Q State covariance (rp x rp)
+# R Observation covariance (n x n)
+# F_0 Initial state vector (rp x 1)
+# P_0 Initial state covariance (rp x rp)
+# retLL compute log likelihood
+
+function SKFS(X::Matrix, A::Matrix, C::Matrix, Q::Matrix,
+              R::Matrix, F_0::Vector, P_0::Matrix, retLL = false)
+
+    T = size(X)[1]
+    n = size(X)[2]
+    rp = size(A)[1]
+    n_c = Int64(0)
+
+    # In internal code factors are Z (instead of F) and factor covariance V (instead of P),
+    # to avoid confusion between the matrices and their predicted (p) and filtered (f) states.
+    # Additionally the results matrices for all time periods have a T in the name.
+
+    loglik = retLL ? 0 : NaN
+    dn = 0
+    if retLL
+        dn = n * log(2.0*pi)
+    end
+    Zf = F_0
+    Vf = P_0
+
+    # Predicted state mean and covariance
+    ZTp = zeros(rp,T)
+    VTp = zeros(rp,rp,T)
+
+    # Filtered state mean and covariance
+    ZTf = zeros(rp,T)
+    VTf = zeros(rp,rp,T)
+
+    # Handling missing values in the filter
+    arow = nmiss = findall(isfinite.(A[1,:]))
+    if isempty(nmiss)
+        error("Missing first row of transition matrix")
+    end
+
+
+    for i in 1:T
+
+        # Run a prediction
+        Zp = A * Zf
+        Vp = A * Vf * transpose(A) + Q
+        Vp += transpose(Vp) # Ensure symmetry
+        Vp *= 0.5
+
+        # If missing observations are present at some timepoints, exclude the
+        # appropriate matrix slices from the filtering procedure.
+        xt = X[i,:] 
+        nmiss = findall(isfinite.(xt))
+        n_c = length(nmiss)
+
+        if n_c > 0
+            if n_c == n
+                Ci = C
+                Ri = R
+            else
+                Ci = C[nmiss,arow]
+                Ri = R[nmiss,nmiss]
+                xt = xt[nmiss]
+            end
+
+            # Intermediate results
+            VCt = Vp * transpose(Ci)
+            S = inv(Symmetric(Ci * VCt + Ri))
+
+            # Prediction error
+            et = xt - Ci * Zp
+            # Kalman gain
+            K = VCt * S
+            # Updated state estimate
+            Zf = Zp + K * et
+            # Updated state covariance estimate
+            Vf = Vp - K * Ci * Vp
+            Vf += transpose(Vf) # Ensure symmetry
+            Vf *= 0.5
+
+            # Compute likelihood. Skip this part if S is not positive definite.
+            if retLL
+                detS = det(S)
+                if detS > 0
+                    loglik += log(detS) - transpose(et) * S * et - dn # Standard Kalman Filter Likelihood
+                end
+            end
+
+        else # If all missing: just prediction.
+            Zf = Zp
+            Vf = Vp
+        end
+
[…]
