### 2023-03-30 37b922c
Not testing this on CRAN, intel MKL does not give stable ordering.
