+++ NEWS.md
@@ -1,8 +1,8 @@
-# dfms 0.2.0.9000
+# dfms 0.2.0
 
 * Added argument `idio.ar1 = TRUE` allowing estimation of approximate DFM's with AR(1) observation errors. 
 
-* Added a small theoretical vignette entitled 'Dynamic Factor Models: A Very Short Introduction'. This vignette lays a foundation for present and future functionality of *dfms*. I plan to implement all features described in this vignette until summer 2023. 
+* Added a small theoretical vignette entitled 'Dynamic Factor Models: A Very Short Introduction'. This vignette lays a foundation for the present and future functionality of *dfms*. I plan to implement all features described in this vignette until summer 2023. 
 
 # dfms 0.1.5
 
