+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: dfms
-Version: 0.2.0.9000
+Version: 0.2.0
 Title: Dynamic Factor Models
 Authors@R: c(person("Sebastian", "Krantz", role = c("aut", "cre"), email = "sebastian.krantz@graduateinstitute.ch"),
              person("Rytis", "Bagdziunas", role = "aut"))
