### 2024-11-10 #556 @emilyriederer
Hi @SebKrantz - checking in with the rOpenSci editors team. Are you still actively working on this? Do you have line of sight to wanting to finish this review?
