### 2025-02-02 #556 @maurolepore
Dear @SebKrantz this is to mark the start of my EiC rotation. I'm reviewing all open issues and noting what I see. Do you intend to continue?

* I see one review has already been submitted [here](https://github.com/ropensci/software-review/issues/556#issuecomment-1369469742).
* @eeholmes shared several comments but the formal review seems pending.
* The author [might be unavailable](https://github.com/ropensci/software-review/issues/556#issuecomment-2467720976), so before insisting about the review it seems important to ensure the author intends to continue.
* The submission has been holding for about 1 year. WARNING: Following our policy (below) the bot may close this issue soon.

---

https://devguide.ropensci.org/softwarereview_policies.html#policiesreviewprocess

> The author can choose to have their submission put on hold (editor applies the holding label). The holding status will be revisited every 3 months, and after one year the issue will be closed.
