### 2025-02-02 #556 @SebKrantz
Dear @maurolepore, thanks! Yes, still planning to do this. Currently resubmitting collapse JSS article and I have a presentation on 10th Feb which I need to prepare. Hope there is some time afterwards.
