### 2025-02-02 #556 @maurolepore
@SebKrantz, thanks for your quick response.  I'll try to reset the holding-clock by removing the tag and re-assigning it. This intends to stop the but from closing this issue just before you're ready to continue.
