### 2025-05-04 #556 @maurolepore
@SebKrantz I'm wrapping up my EiC rotation so I'm leaving a few notes to hand over this submission to the next EiC:

* I summarized the situation 3 months ago [here](https://github.com/ropensci/software-review/issues/556#issuecomment-2629505671).
* Since then I reset the holding clock to avoid closing because the hold remained for 1 year but the @SebKrantz intended to resume right about then.
* I see no more activity so we need to confirm the intention to continue or close.
