### 2025-05-11 #556 @SebKrantz
My response to the review by @santikka below. Will look into the one by @eeholmes during the week...

> ## Package Review
> I'm currently unable to assess how well this package conforms to the standards for two reasons. First, all of the `@srrstats` tags are still included in the `srr-stats-standards.R` file and are not placed in appropriate places in the code. Second, the compliance of only a handful of the standards has been explained.

- I've placed (blocks of) them in the files now. I've provided some commentary under some of them, but in general: Inputs to DFM() are now rigorously checked, the package is robust to noise (its a factor model after all), supports missing values, but cannot be used with character or complex data and will likely not converge with infinite values.

> #### Algorithms
>
> As I understand it, the one of the main advantages of this package is computational speed, making C++ a valid choice for implementing the core algorithms. I would like to see some tests to substantiate the performance claims of this package vs. packages such as `MARSS` and `bayesdfa` that were mentioned by the authors.

- Of course such tests would be nice, but I'm not sure the model of Banbura & Modugno (2014) can even be estimated with MARSS. In any case, MARSS is implemented fully in R, so the C++-based Kalman filter alone should give a significant speedup. Bayesian methods are also more expensive, but indeed here some tests would be interesting.

> #### Testing
>
> While the package has impressive test coverage, I found the suite of tests lacking especially with regards to the inputs for the exported functions. There are several instances where the inputs are not thoroughly checked, for example:
>
> .VAR(diff(EuStockMarkets), -1)
>  Error in `[.default`(x, (p + 1L - i):(TT - i), ) :
> only 0's may be mixed with negative subscripts
>

- Thanks, I've added a condition to check that p is positive now.

> I'm also not convinced that the results of `autotest` have been properly accounted for. At the time of writing this review, running `autotest` on the package produces a data frame with 200 rows.

- TBH I have a problem understanding the results of this package. Take for example the Arguments rR and rQ in DFM() specifying the restrictions places on the observation and transition covariance matrix. They are checked as follows:

```r
# Default values
rQ = c("none", "diagonal", "identity")
rR = c("diagonal", "identity", "none")

# Input Checks
rRi <- switch(tolower(rR[1L]), identity = 0L, diagonal = 1L, none = 2L, stop("Unknown rR option:", rR[1L]))
rQi <- switch(tolower(rQ[1L]), identity = 0L, diagonal = 1L, none = 2L, stop("Unknown rQ option:", rQ[1L]))
```

Now autotest gives:

+-------+------------------------+---------+-----------+------------------+----------------------------------------+-------------------------------------------+------+----------------------------------+
| type  | test_name              | fn_name | parameter | parameter_type   | operation                              | content                                   | test | yaml_hash                        |
+=======+========================+=========+===========+==================+========================================+===========================================+======+==================================+
| dummy | single_char_case       | DFM     | rQ        | single character | lower-case character parameter         | (Should yield same result)                | TRUE | 960e3636be208d42bd2c7781cc3f322e |
+-------+------------------------+---------+-----------+------------------+----------------------------------------+-------------------------------------------+------+----------------------------------+
| dummy | single_char_case       | DFM     | rQ        | single character | upper-case character parameter         | (Should yield same result)                | TRUE | 960e3636be208d42bd2c7781cc3f322e |
+-------+------------------------+---------+-----------+------------------+----------------------------------------+-------------------------------------------+------+----------------------------------+
| dummy | random_char_string     | DFM     | rQ        | single character | random character string as parameter   | Should error                              | TRUE | 960e3636be208d42bd2c7781cc3f322e |
+-------+------------------------+---------+-----------+------------------+----------------------------------------+-------------------------------------------+------+----------------------------------+
| dummy | single_par_as_length_2 | DFM     | rQ        | single character | Length 2 vector for length 1 parameter | Should trigger message, warning, or error | TRUE | 960e3636be208d42bd2c7781cc3f322e |
+-------+------------------------+---------+-----------+------------------+----------------------------------------+-------------------------------------------+------+----------------------------------+
| dummy | single_char_case       | DFM     | rR        | single character | lower-case character parameter         | (Should yield same result)                | TRUE | 960e3636be208d42bd2c7781cc3f322e |
+-------+------------------------+---------+-----------+------------------+----------------------------------------+-------------------------------------------+------+----------------------------------+
| dummy | single_char_case       | DFM     | rR        | single character | upper-case character parameter         | (Should yield same result)                | TRUE | 960e3636be208d42bd2c7781cc3f322e |
+-------+------------------------+---------+-----------+------------------+----------------------------------------+-------------------------------------------+------+----------------------------------+
| dummy | random_char_string     | DFM     | rR        | single character | random character string as parameter   | Should error                              | TRUE | 960e3636be208d42bd2c7781cc3f322e |
+-------+------------------------+---------+-----------+------------------+----------------------------------------+-------------------------------------------+------+----------------------------------+
| dummy | single_par_as_length_2 | DFM     | rR        | single character | Length 2 vector for length 1 parameter | Should trigger message, warning, or error | TRUE | 960e3636be208d42bd2c7781cc3f322e |
+-------+------------------------+---------+-----------+------------------+----------------------------------------+-------------------------------------------+------+----------------------------------+

Why? I fail to understand the problem.

> #### Visualisation (where appropriate)
>
> Visualisations are mostly clear and aid the primary purposes of statistical interpretation of the results. I don't think there is a risk of statistical misinterpretation.
>
> There is a small issue where the plot legend may overlap with the plotted values (for example when running `plot(mod, type = "individual", method = "all")` in the examples of `plot.dfm`), perhaps the authors could move the legend outside of the main figure area.

- Thanks, this is a good suggestion. Ideal would be something like legend.position = "top" as in ggplot2 and then put it below the plot title. Not sure though how to implement that reliably using legend() in base R.

> #### Package Design
>
> Overall, the package seems mostly well designed for its intended purpose. The code has been thoroughly annotated making it easy to understand, although some of the lines are very long, and the indentation is sometimes inconsistent.
>
> There are some curious design choices, like using `switch` instead of `match.arg` to check function arguments. Also, this part in `DFM.R` puzzled me:

- switch is much more efficient. Probably not needed here, but just habitual from developing collapse where every microsecond counts.

> # Quoting some functions that need to be evaluated iteratively
> .EM_DGR <- quote(EMstepDGR(X, A, C, Q, R, F_0, P_0, cpX, n, r, sr, TT, rQi, rRi))
> .EM_BM <- quote(EMstepBMOPT(X, A, C, Q, R, F_0, P_0, XW0, W, n, r, sr, TT, dgind, dnkron, dnkron_ind, rQi, rRi))
> .KFS <- quote(SKFS(X, A, C, Q, R, F_0, P_0))
>
> These expressions are then `eval`'d later, why exactly is this done as opposed to simply calling these functions directly?

- So the reason for doing this is that they are iteratively evaluated:

```r
  em_res <- list()
  encl <- environment()
  while(num_iter < max.iter && !converged) {

    em_res <- eval(expr, em_res, encl)
    loglik <- em_res$loglik

    ## Iterate at least min.iter times
    if(num_iter < min.iter) converged <- FALSE else {
      converged <- em_converged(loglik, previous_loglik, tol, check.increased)
      if(check.increased) converged <- converged[1L] && !converged[2L]
    }

    previous_loglik <- loglik
    num_iter <- num_iter + 1L
    loglik_all[num_iter] <- loglik

  }
```
where `expr` is the function being called, e.g. `quote(EMstepBMOPT(X, A, C, Q, R, F_0, P_0, XW0, W, n, r, sr, TT, dgind, dnkron, dnkron_ind, rQi, rRi))`. Some of the function inputs are fixed: `XW0, W, n, r, sr, TT, dgind, dnkron, dnkron_ind, rQi, rRi`, the others are updated every iteration until convergence. I Don't know of a more parsimonious way to code this than quoting a function which is then evaluated against its own results object and the enclosing environment which contains the invariant objects and the initial estimates of the system matrices.

> I'm also curious as to what exactly is the logic in choosing which parts of the code are written in R and which in C++. It seems to me that there is quite a lot of matrix algebra going on in the R code. I would have expected that R would only be used to check and prepare the inputs for C++, and to provide outputs such as plots.

- Agreed, I've been lazy. The entire EM algorithm could also be written in C++. For now its mostly the Kalman Filter and Smoother. C++ is more tedious to write, so I've refrained from that until now. If I get more time on the package I'll rewrite some stuff in C++.

> Print methods should return their respective argument objects invisibly (i.e., for `dfm`, `dfm_summary`, `dfm_forecast`, and `ICr`).

- Thanks, I've done this.

> Since this is a non-tidyverse package, it is limited in terms of inter-operability in relation to other packages. For example, only base R graphics are available.

- As far as ggplots are concerned I would agree, but then it depends on what graphics system the user prefers. I believe the `as.data.frame.dfm` and `as.data.frame.dfm_forecast` methods which return tidy data frames make it quite interoperable. In general, I wanted to create a low-dependency and fast package which warranted a non-tidyverse approach.

>     * [ ]  **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines
>
>
> This package does not conform to the guidelines in at least the following aspects:
>
>     * Inconsistent use of assignment operators

- It turns out that `=` is slightly faster than `<-`, which is why it is used in the the EM algorithms which are called iteratively.

>     * No top-level documentation

- Thanks, I added a package page now.

>     * No CITATION file
>
>     * README should also show the results of the usage example (I suggest using .Rmd)

- Thanks, I executed the README examples. There is currently no dedicated article, so a CITATION file is presently not necessary.

>
> Estimated hours spent reviewing: 5
>
>     * [x]  Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

- Thanks for investing the time! I added you as 'rev' to the description.
