### 2025-05-17 #556 @SebKrantz
And my responses to the review by @eeholmes:

> ## Missing info on the description
>
> https://sebkrantz.github.io/dfms/ and description on CRAN is missing key information that would allow a statistician who is not in econometrics to understand what this does. It would be good to add the following detail:
>
> "It is a DFA where the factors follow a stationary VAR process of order p."
>
> In my field, DFAs are random walks (A=identity) so I was puzzled as to why the A matrix was not identity and why the identifiability constraint was different. Then I read Banbura, M., & Modugno, M. and realized this a different type of DFA. You might also add that this type of DFA is (widely?) used in econometrics.

- Thanks, I added "Factors are assumed to follow a stationary VAR process of order p" to the DESCRIPTION and README. There is now also a new [theoretical vignette](https://raw.githubusercontent.com/SebKrantz/dfms/main/vignettes/dynamic_factor_models_paper.pdf) explaining the model in greater detail.

> ## Minor
>
> -  Readme and vignettes code examples use "=" but R style uses "<-" for assignment.

- Thanks, fixed.

> -  https://sebkrantz.github.io/dfms/reference/DFM.html says it returns a class dfm which is "list-like". like? I looks like a list not "list-like". If it is a list then set class to c("dfm", "list") so that list methods recognize it as a list.

- This is interesting, but I don't see base R `lm()` etc. append the list class to its result object. I am hesitant to do something that is not common practice.

> -  `?dfms` did not pull up the help page. Is there no help page for the package?

- I added it now.

> -  https://sebkrantz.github.io/dfms/reference/SKFS.html
> The covariance equation has a typo. I think you meant F^{smooth}_t and left out the { }

- Thanks, fixed.

> ## Kalman filter is not only for stationary data
>
> https://sebkrantz.github.io/dfms/reference/SKF.html
> https://sebkrantz.github.io/dfms/reference/SKFS.html
>
> The Kalman filter and smoother are not restricted to stationary data. A can be identity (random walks).
>
> You have to know a lot about state-space models and the Kalman filter and smoother to understand the documentation. However it does cite the references for the user.

- Thanks. I agree that I didn't consider the Random Walk case. What I meant here is that the Kalman Filter does not model trends or intercepts in the data. The naming convention SKF = Stationary Kalman Filter follows Banbura & Modugno (2014), and I don't want to change function names now. In response to this comment, I added a qualifier: "(or random walks - data should be mean zero and without a trend)" to the Description.

> ## Add mention of the data assumptions
>
> The model is stationary but no stationarity tests for the data or mention of the importance of this. Also the data must have 0 mean due to no mean in the model.
>
> I would make this quite clear for the user and point them towards tests for stationarity.

- Thanks. The data must not have mean zero because it is scaled and centered in `DFM()`. I have made this more explicit in the documentation now. I don't plan to provide stationarity tests in this package, and as you mentioned, it would also work with a random walk. Estimation works even with mild trends, noting of course that the functional form doesn't reflect this. I also don't provide inference procedures for the factors (yet), so testing is not very critical - eyeballing the data should suffice. In economics there are certain conventions, e.g. interest/inflation rates in levels or differences, and all other variables in log-differences --- as shown in the examples. This generally produces stationary data.

> ## no coef() and logLik() functions
>
> Not big deal but I would expect those for a statistics package.

- Thanks. I added them now.

> ## class for DFA output
>
> The class is shown as "dfm", but it appears to be a list also. In this case, one should prepend "dfm" but keep "list" also so that list methods work.

- Thanks. However, I don't know of any major model class where this is the case. It is also not the case for data.frame. The default method usually gives the correct behavior on the list-vector.

> ## The method of imputation for X_imp is not given
>
> https://sebkrantz.github.io/dfms/reference/DFM.html
>
> In the value section, X_imp is one of the values returned, but how it is imputed is not given. Is this the Kalman smoother output from the estimated model? Or something else? tsnarmimp?

- Thanks. I have given some details. Basically, this is `tsnarmimp(scale(X), ...)`.

> ## No info on the residuals
>
> No information on what type of residuals are returned. Link is to the vars residuals() function, but that has no information either. I am assuming that the economics DFA field, there are a standard type of residuals being used, but would be helpful to say which ones. *Note, I'll add the options later. Need to look them up.*

- Thanks, `?resid.dfm` now states that the residuals are computed as `x_t - C*F_t`, where `C` is the observation matrix.

> ## No info on fitted.
>
> Same re fitted(). No info in the vars help page either. For conditional state-space models, the meaning of "fitted" can have multiple definitions. I can make a good guess, but it'd be better to be explicit by saying. *I'll add an example later*

- Same here, it is `C*F_t`, where `C` is the observation matrix - stated in the Description now.

> ## Nothing mentioned about the F_0 and P_0 assumptions
>
> Different algorithms and authors take different approaches. Which one is used here? *I'll need to read Banbura, M., & Modugno, M.  to see what they do.*

- I added the sentence: "The filter is initialized with PCA estimates on the imputed dataset—see ?SKFS for a complete example." In short, `F_0` is the first row of the PCA estimates. P_0 is `inv(I - kron(A, A)) * Q`.

> ## What's going on with the rotation?
>
> Why is C not returned with the identifiability constraint? For my test case, my data has missing values so Banbura, M., & Modugno, M. should have been used. They specify a identifiability contraint with the $r \times r$ matrix at top of C set to identity. But the returned C had all values. So that means there is a rotation being done post estimation. Need to look that up and that info should be added to the DFM() details.

- Not exactly sure what the problem here is. Let me just make one argument. Doz et al. (2012) argue that as the number of `n` tends to infinity with the number of periods `T` kept fixed, the dynamic factors become identical to PCA estimates. But principal components are orthogonal only to each other, not to the data. In this case C resembles the loadings matrix (eigenvectors) from PCA, which is a dense matrix. And yes, in PCA we can rotate the PC's to better match certain characteristics (Varimax etc.). `DFM()` does no such thing, just the default PC's are used to initialize to he Kalman filter and the final factors generally follow the same pattern.

> ## Installation
>
> Installation failed on my M1 Mac at first. Probably just my machine set-up's fault.
>
> ```
> ld: warning: directory not found for option '-L/opt/R/arm64/gfortran/lib/gcc/aarch64-apple-darwin20.6.0/12.0.1'
> ld: warning: directory not found for option '-L/opt/R/arm64/gfortran/lib'
> ld: library not found for -lgfortran
> clang: error: linker command failed with exit code 1 (use -v to see invocation)
> make: *** [dfms.so] Error 1
> ERROR: compilation failed for package ‘dfms’
> ```
>
> Problem: no fortran compiler for M1 Macs in XCode. Problem discussed here: https://mac.r-project.org/tools/
>
> Solution posted here https://github.com/RubD/Giotto_site/issues/11
> ```
> # for R>=4.2.0
> curl -O https://mac.r-project.org/tools/gfortran-12.0.1-20220312-is-darwin20-arm64.tar.xz
>
> # unpack
> sudo tar fxz gfortran-12.0.1-20220312-is-darwin20-arm64.tar.xz -C /
>
> # /opt/R/arm64/gfortran/SDK has to point to your macOS SDK
> # EEH: this didn't work for me but was able to install anyhow and run example
> sudo gfortran-update-sdk
> ```

- Thanks, this seems like an old Rcpp or RcppArmadillo issue. *dfms* has no Fortran source code.

Overall, highly appreciate the effort even though this econometrics literature, which I closely follow with this package, may not be much of a pure statistical treatment of DFMs. But it works and is computationally efficient. I also added you as 'rev' to the description now.
