### 2025-05-20 #556 @mpadge
Great, thanks @SebKrantz. Now back to both of you @eeholmes @santikka Could you both:

- Have a look over @SebKrantz's comments both [here](https://github.com/ropensci/software-review/issues/556#issuecomment-2870265446) and [here](https://github.com/ropensci/software-review/issues/556#issuecomment-2888618676),
- Compare comments with package updates to confirm all issues have been satisfactorily addressed.
- Feel free to comment here about any issues that may remain, and then once everything has finally been resolved:
- Paste the [reviewer approval template](https://devguide.ropensci.org/approval2template.html) in a separate comment, fill out the details, and click "Comment" here.

Thanks in advance, and thank you @SebKrantz for getting this longstanding review into motion again 🚀
