### 2026-01-22 #556 @jhollist
@SebKrantz, apologies for the long delay in getting this through the process.  Think it was just one of those things were schedules were misaligned and it fell off our radar!

I have had some discussions with the editorial team and we are happy to say we can move this forward.  Your response to the revisions were thorough and complete.  I will officially accept this shortly.

There will be info in this issue on what needs to happen next once I do that. If you have any questions or issues, just holler.

Thanks for your contribution and thank you to all for your reviews and support of rOpenSci!
